refactor(node): move node assembly in to zone-node crate
