+++ crates/node/Cargo.toml
@@ -1,0 +1,91 @@
+[package]
+name = "zone-node"
+description = "Tempo Zone node assembly"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# tempo
+tempo-chainspec.workspace = true
+tempo-evm = { workspace = true, features = ["rpc", "engine"] }
+tempo-node.workspace = true
+tempo-primitives.workspace = true
+tempo-transaction-pool.workspace = true
+tempo-alloy.workspace = true
+tempo-contracts.workspace = true
+tempo-zone-contracts = { workspace = true, features = ["std", "serde", "rpc"] }
+zone-evm.workspace = true
+zone-l1.workspace = true
+zone-payload.workspace = true
+zone-primitives = { workspace = true, features = ["std", "serde"] }
+zone-rpc.workspace = true
+zone-sequencer.workspace = true
+
+# reth
+reth-chainspec.workspace = true
+reth-consensus = { workspace = true, optional = true }
+reth-eth-wire-types.workspace = true
+reth-ethereum = { workspace = true, features = [
+  "node",
+  "cli",
+], optional = true }
+reth-node-api.workspace = true
+reth-node-builder.workspace = true
+reth-payload-builder.workspace = true
+reth-payload-primitives.workspace = true
+reth-primitives-traits.workspace = true
+reth-provider.workspace = true
+reth-rpc.workspace = true
+reth-rpc-builder.workspace = true
+reth-rpc-eth-api.workspace = true
+reth-rpc-eth-types.workspace = true
+reth-storage-api.workspace = true
+reth-tasks.workspace = true
+reth-tracing = { workspace = true, optional = true }
+reth-transaction-pool.workspace = true
+
+# alloy
+alloy-consensus.workspace = true
+alloy-network.workspace = true
+alloy-primitives.workspace = true
+alloy-provider.workspace = true
+alloy-rpc-client.workspace = true
+alloy-rpc-types-engine.workspace = true
+alloy-rpc-types-eth.workspace = true
+alloy-signer-local.workspace = true
+alloy-sol-types.workspace = true
+
+# misc
+clap = { workspace = true, optional = true }
+eyre.workspace = true
+futures.workspace = true
+k256.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[features]
+default = []
+cli = [
+  "dep:clap",
+  "dep:reth-consensus",
+  "dep:reth-ethereum",
+  "dep:reth-tracing",
+  "tempo-chainspec/cli",
+]
+jemalloc = ["reth-ethereum?/jemalloc"]
+test-utils = [
+  "reth-chainspec/test-utils",
+  "reth-ethereum?/test-utils",
+  "reth-node-builder/test-utils",
+  "reth-payload-builder/test-utils",
+  "reth-primitives-traits/test-utils",
+  "reth-provider/test-utils",
+  "reth-transaction-pool/test-utils",
+  "tempo-transaction-pool/test-utils",
+]
