+++ crates/tempo-zone/Cargo.toml
@@ -12,64 +12,15 @@ workspace = true
 
 [dependencies]
 # tempo
-tempo-chainspec.workspace = true
-tempo-evm = { workspace = true, features = ["rpc", "engine"] }
-tempo-node.workspace = true
-tempo-primitives.workspace = true
-tempo-transaction-pool.workspace = true
-tempo-alloy.workspace = true
-tempo-contracts.workspace = true
 tempo-zone-contracts = { workspace = true, features = ["std", "serde", "rpc"] }
 zone-evm.workspace = true
 zone-l1.workspace = true
+zone-node.workspace = true
 zone-payload.workspace = true
-zone-primitives = { workspace = true, features = ["std", "serde"] }
-zone-rpc.workspace = true
 zone-sequencer.workspace = true
 
-# reth
-reth-chainspec.workspace = true
-reth-eth-wire-types.workspace = true
-reth-evm.workspace = true
-reth-metrics.workspace = true
-reth-node-api.workspace = true
-reth-node-builder.workspace = true
-reth-payload-builder.workspace = true
-reth-payload-primitives.workspace = true
-reth-primitives-traits.workspace = true
-reth-provider.workspace = true
-reth-rpc.workspace = true
-reth-rpc-builder.workspace = true
-reth-rpc-eth-api.workspace = true
-reth-rpc-eth-types.workspace = true
-reth-storage-api.workspace = true
-reth-tasks.workspace = true
-reth-transaction-pool.workspace = true
-
-# alloy
-alloy-consensus.workspace = true
-alloy-network.workspace = true
-alloy-primitives.workspace = true
-alloy-provider.workspace = true
-alloy-rpc-client.workspace = true
-alloy-rpc-types-engine.workspace = true
-alloy-rpc-types-eth.workspace = true
-alloy-signer-local.workspace = true
-alloy-sol-types.workspace = true
-
 # misc
-clap = { workspace = true, optional = true }
-k256.workspace = true
-reth-consensus = { workspace = true, optional = true }
-reth-ethereum = { workspace = true, features = [
-  "node",
-  "cli",
-], optional = true }
-reth-tracing = { workspace = true, optional = true }
 eyre.workspace = true
-futures.workspace = true
-tokio.workspace = true
-tracing.workspace = true
 
 [dev-dependencies]
 alloy = { workspace = true, features = [
@@ -81,21 +32,41 @@ alloy = { workspace = true, features = [
   "signer-mnemonic-all-languages",
   "rpc-types",
 ] }
+alloy-consensus.workspace = true
 alloy-contract.workspace = true
 alloy-eips.workspace = true
 alloy-evm.workspace = true
+alloy-network.workspace = true
 alloy-primitives.workspace = true
+alloy-provider.workspace = true
 alloy-rlp.workspace = true
+alloy-rpc-client.workspace = true
 alloy-rpc-types-engine.workspace = true
+alloy-rpc-types-eth.workspace = true
+alloy-signer-local.workspace = true
 alloy-sol-types.workspace = true
 base64.workspace = true
 const-hex.workspace = true
+futures.workspace = true
+k256.workspace = true
 p256.workspace = true
 rand.workspace = true
+reth-chainspec.workspace = true
+reth-eth-wire-types.workspace = true
 reth-ethereum = { workspace = true, features = ["node", "test-utils"] }
+reth-node-api.workspace = true
 reth-node-builder.workspace = true
 reth-node-core.workspace = true
+reth-payload-builder.workspace = true
+reth-payload-primitives.workspace = true
+reth-primitives-traits.workspace = true
+reth-provider.workspace = true
 reth-rpc-builder.workspace = true
+reth-rpc-eth-api.workspace = true
+reth-rpc-eth-types.workspace = true
+reth-storage-api.workspace = true
+reth-tasks.workspace = true
+reth-transaction-pool.workspace = true
 reth-tracing.workspace = true
 serde = { workspace = true, features = ["derive"] }
 alloy-signer.workspace = true
@@ -103,34 +74,28 @@ reqwest.workspace = true
 revm.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
+tempo-alloy.workspace = true
 tempo-chainspec.workspace = true
 tempo-contracts.workspace = true
+tempo-evm = { workspace = true, features = ["rpc", "engine"] }
+tempo-node.workspace = true
 tempo-payload-types.workspace = true
+tempo-primitives.workspace = true
 tempo-revm.workspace = true
+tempo-transaction-pool.workspace = true
 sha2.workspace = true
 tempo-precompiles = { workspace = true, features = ["test-utils"] }
 tokio = { workspace = true, features = ["rt", "rt-multi-thread", "macros"] }
 tokio-tungstenite.workspace = true
+tracing.workspace = true
 url = "2"
+zone-primitives = { workspace = true, features = ["std", "serde"] }
 
 [features]
 default = []
-cli = [
-  "dep:clap",
-  "dep:reth-consensus",
-  "dep:reth-ethereum",
-  "dep:reth-tracing",
-  "tempo-chainspec/cli",
-]
-jemalloc = ["reth-ethereum?/jemalloc"]
+cli = ["zone-node/cli"]
+jemalloc = ["zone-node/jemalloc"]
 test-utils = [
-  "reth-chainspec/test-utils",
-  "reth-ethereum/test-utils",
-  "reth-node-builder/test-utils",
-  "reth-payload-builder/test-utils",
-  "reth-primitives-traits/test-utils",
-  "reth-provider/test-utils",
-  "reth-transaction-pool/test-utils",
+  "zone-node/test-utils",
   "tempo-precompiles/test-utils",
-  "tempo-transaction-pool/test-utils",
 ]
