+++ crates/tempo-zone/src/lib.rs
@@ -4,8 +4,6 @@
 #![allow(unnameable_types)]
 #![allow(clippy::too_many_arguments)]
 use eyre as _;
-use reth_evm as _;
-use reth_metrics as _;
 
 pub mod abi;
 #[cfg(feature = "cli")]
